requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>

tensorflow-gpu=2.4.1
scikit-image=0.16.2
scikit-learn=0.24.2
pandas=1.2.5
pillow=8.3.1
matplotlib=3.1.2
tqdm=4.62.3
ujson=4.0.2
tifffile=2021.7.2
tensorflow-addons=0.13.0